simplify the consensus graph generation

+++ src/consensus_graph.cpp
@@ -759,6 +759,7 @@ odgi::graph_t* create_consensus_graph(const xg::XG &smoothed,
     }
 
     /// TODO validate consensus graph until here
+
     // FIXME: this should check the actual path sequence for validation
     // not each step
     /*
@@ -802,7 +803,8 @@ odgi::graph_t* create_consensus_graph(const xg::XG &smoothed,
             });
     */
 
-    consensus->for_each_path_handle(
+    // TODO this does not seem to be necessary
+    /*consensus->for_each_path_handle(
         [&](const path_handle_t& path) {
             consensus->for_each_step_in_path(path, [&] (const step_handle_t step) {
                if (consensus->has_next_step(step)) {
@@ -814,11 +816,10 @@ odgi::graph_t* create_consensus_graph(const xg::XG &smoothed,
                    }
                }
             });
-        });
+        });*/
 
-    // unchop the graph
-    // this unchop is necessary
-    odgi::algorithms::unchop(*consensus, thread_count, false);
+    // This is necessary
+    odgi::algorithms::unchop(*consensus, thread_count, true);
 
     std::cerr << "[smoothxg::create_consensus_graph] removing edges connecting the path with a gap less than " << min_allele_length << "bp" << std::endl;
 
@@ -916,7 +917,6 @@ odgi::graph_t* create_consensus_graph(const xg::XG &smoothed,
     // TODO this does not seem to be necessary
     /*
     consensus->for_each_path_handle(
-
         [&](const path_handle_t& path) {
             consensus->for_each_step_in_path(path, [&] (const step_handle_t step) {
                if (consensus->has_next_step(step)) {
@@ -931,22 +931,25 @@ odgi::graph_t* create_consensus_graph(const xg::XG &smoothed,
         });
         */
 
-    odgi::algorithms::unchop(*consensus, thread_count, false);
+    // This is necessary
+    odgi::algorithms::unchop(*consensus, thread_count, true);
 
     auto* copy = new odgi::graph_t();
     graph_deep_copy(consensus, copy);
     delete consensus;
     consensus = copy;
 
-    odgi::algorithms::unchop(*consensus, thread_count, false);
+    //odgi::algorithms::unchop(*consensus, thread_count, true);
 
     // remove 0-depth nodes and edges
-    auto handles_to_drop = odgi::algorithms::find_handles_exceeding_coverage_limits(*consensus, 1, 0);
-    for (auto& handle : handles_to_drop) {
-        consensus->destroy_handle(handle);
-    }
+    std::vector<handle_t> handles_to_drop = odgi::algorithms::find_handles_exceeding_coverage_limits(*consensus, 1, 0);
+    if (!handles_to_drop.empty()) {
+        for (auto& handle : handles_to_drop) {
+            consensus->destroy_handle(handle);
+        }
 
-    odgi::algorithms::unchop(*consensus, thread_count, false);
+        odgi::algorithms::unchop(*consensus, thread_count, true);
+    }
 
     uint64_t consensus_nodes = 0;
     uint64_t consensus_length = 0;
